add: - 결제 기능   - 결제 요청 구현     - 요청시 로컬 DB에 요청기록 저장   - 결제 처리 server에 응답 받음     - json 응답값까지 받아놓음 - ignore   - apikey 무시
fix :
- 파일구조 수정

+++ src/main/java/com/example/UserMapper.java
@@ -7,6 +7,6 @@ import org.apache.ibatis.annotations.Select;
 @Mapper
 public interface UserMapper {
 
-    User getUserByIdTest(@Param("id") Long id);
+    User getUserByIdTest(@Param("id") int id);
 
 }
